Move software installer tests to enterprise, not mdm
