+++ server/service/integration_mdm_test.go
@@ -6,7 +6,6 @@ import (
 	"crypto/x509"
 	"database/sql"
 	"encoding/base64"
-	"encoding/hex"
 	"encoding/json"
 	"encoding/xml"
 	"errors"
@@ -361,12 +360,6 @@ func (s *integrationMDMTestSuite) TearDownTest() {
 	})
 	mysql.ExecAdhocSQL(t, s.ds, func(tx sqlx.ExtContext) error {
 		_, err := tx.ExecContext(ctx, "DELETE FROM host_mdm_apple_declarations")
-		return err
-	})
-
-	// clear any lingering software installers
-	mysql.ExecAdhocSQL(t, s.ds, func(tx sqlx.ExtContext) error {
-		_, err := tx.ExecContext(ctx, "DELETE FROM software_installers")
 		return err
 	})
 }
@@ -8465,614 +8458,4 @@ func (s *integrationMDMTestSuite) TestIsServerBitlockerStatus() {
 	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/hosts/%d", host2.ID), nil, http.StatusOK, &hr)
 	require.NotNil(t, hr.Host.MDM.OSSettings.DiskEncryption.Status)
 	require.Equal(t, fleet.DiskEncryptionEnforcing, *hr.Host.MDM.OSSettings.DiskEncryption.Status)
-}
-
-func (s *integrationMDMTestSuite) TestSoftwareInstallerUploadDownloadAndDelete() {
-	t := s.T()
-
-	openFile := func(name string) *os.File {
-		f, err := os.Open(filepath.Join("testdata", "software-installers", name))
-		require.NoError(t, err)
-		return f
-	}
-
-	var expectBytes []byte
-	var expectLen int
-	f := openFile("ruby.deb")
-	st, err := f.Stat()
-	require.NoError(t, err)
-	expectLen = int(st.Size())
-	require.Equal(t, expectLen, 11340)
-	expectBytes = make([]byte, expectLen)
-	n, err := f.Read(expectBytes)
-	require.NoError(t, err)
-	require.Equal(t, n, expectLen)
-	f.Close()
-
-	checkDownloadResponse := func(t *testing.T, r *http.Response, expectedFilename string) {
-		require.Equal(t, "application/octet-stream", r.Header.Get("Content-Type"))
-		require.Equal(t, fmt.Sprintf(`attachment;filename="%s"`, expectedFilename), r.Header.Get("Content-Disposition"))
-		require.NotZero(t, r.ContentLength)
-		require.Equal(t, expectLen, int(r.ContentLength))
-		b, err := io.ReadAll(r.Body)
-		require.NoError(t, err)
-		require.Equal(t, expectLen, len(b))
-		require.Equal(t, expectBytes, b)
-	}
-
-	checkSoftwareTitle := func(t *testing.T, title string, source string) uint {
-		var id uint
-		mysql.ExecAdhocSQL(t, s.ds, func(q sqlx.ExtContext) error {
-			return sqlx.GetContext(context.Background(), q, &id, `SELECT id FROM software_titles WHERE name = ? AND source = ? AND browser = ''`, title, source)
-		})
-		return id
-	}
-
-	checkScriptContentsID := func(t *testing.T, id uint, expectedContents string) {
-		var contents string
-		mysql.ExecAdhocSQL(t, s.ds, func(q sqlx.ExtContext) error {
-			return sqlx.GetContext(context.Background(), q, &contents, `SELECT contents FROM script_contents WHERE id = ?`, id)
-		})
-		require.Equal(t, expectedContents, contents)
-	}
-
-	checkSoftwareInstaller := func(t *testing.T, payload *fleet.UploadSoftwareInstallerPayload) (installerID uint, titleID uint) {
-		var id uint
-		mysql.ExecAdhocSQL(t, s.ds, func(q sqlx.ExtContext) error {
-			var tid uint
-			if payload.TeamID != nil {
-				tid = *payload.TeamID
-			}
-			return sqlx.GetContext(context.Background(), q, &id, `SELECT id FROM software_installers WHERE global_or_team_id = ? AND filename = ?`, tid, payload.Filename)
-		})
-		require.NotZero(t, id)
-
-		meta, err := s.ds.GetSoftwareInstallerMetadata(context.Background(), id)
-		require.NoError(t, err)
-
-		if payload.TeamID != nil {
-			require.Equal(t, *payload.TeamID, *meta.TeamID)
-		} else {
-			require.Nil(t, meta.TeamID)
-		}
-
-		checkScriptContentsID(t, meta.InstallScriptContentID, payload.InstallScript)
-
-		if payload.PostInstallScript != "" {
-			require.NotNil(t, meta.PostInstallScriptContentID)
-			checkScriptContentsID(t, *meta.PostInstallScriptContentID, payload.PostInstallScript)
-		} else {
-			require.Nil(t, meta.PostInstallScriptContentID)
-		}
-
-		require.Equal(t, payload.PreInstallQuery, meta.PreInstallQuery)
-		require.Equal(t, payload.StorageID, meta.StorageID)
-		require.Equal(t, payload.Filename, meta.Name)
-		require.Equal(t, payload.Version, meta.Version)
-		require.Equal(t, checkSoftwareTitle(t, payload.Title, "deb_packages"), *meta.TitleID)
-		require.NotZero(t, meta.UploadedAt)
-
-		return meta.InstallerID, *meta.TitleID
-	}
-
-	t.Run("upload no team software installer", func(t *testing.T) {
-		payload := &fleet.UploadSoftwareInstallerPayload{
-			InstallScript:     "some install script",
-			PreInstallQuery:   "some pre install query",
-			PostInstallScript: "some post install script",
-			Filename:          "ruby.deb",
-			// additional fields below are pre-populated so we can re-use the payload later for the test assertions
-			Title:     "ruby",
-			Version:   "1:2.5.1",
-			Source:    "deb_packages",
-			StorageID: "df06d9ce9e2090d9cb2e8cd1f4d7754a803dc452bf93e3204e3acd3b95508628",
-		}
-
-		s.uploadSoftwareInstaller(payload, http.StatusOK, "")
-
-		// check activity
-		s.lastActivityOfTypeMatches(fleet.ActivityTypeAddedSoftware{}.ActivityName(), `{"software_title": "ruby", "software_package": "ruby.deb", "team_name": null, "team_id": null}`, 0)
-
-		// check the software installer
-		_, titleID := checkSoftwareInstaller(t, payload)
-
-		// upload again fails
-		s.uploadSoftwareInstaller(payload, http.StatusConflict, "already exists")
-
-		// download the installer
-		s.Do("GET", fmt.Sprintf("/api/latest/fleet/software/%d/package?alt=media", titleID), nil, http.StatusBadRequest)
-
-		// delete the installer
-		s.Do("DELETE", fmt.Sprintf("/api/latest/fleet/software/%d/package", titleID), nil, http.StatusBadRequest)
-	})
-
-	t.Run("create team software installer", func(t *testing.T) {
-		var createTeamResp teamResponse
-		s.DoJSON("POST", "/api/latest/fleet/teams", &fleet.Team{
-			Name: t.Name(),
-		}, http.StatusOK, &createTeamResp)
-		require.NotZero(t, createTeamResp.Team.ID)
-
-		payload := &fleet.UploadSoftwareInstallerPayload{
-			TeamID:            &createTeamResp.Team.ID,
-			InstallScript:     "another install script",
-			PreInstallQuery:   "another pre install query",
-			PostInstallScript: "another post install script",
-			Filename:          "ruby.deb",
-			// additional fields below are pre-populated so we can re-use the payload later for the test assertions
-			Title:     "ruby",
-			Version:   "1:2.5.1",
-			Source:    "deb_packages",
-			StorageID: "df06d9ce9e2090d9cb2e8cd1f4d7754a803dc452bf93e3204e3acd3b95508628",
-		}
-
-		s.uploadSoftwareInstaller(payload, http.StatusOK, "")
-
-		// check the software installer
-		installerID, titleID := checkSoftwareInstaller(t, payload)
-
-		// check activity
-		s.lastActivityOfTypeMatches(fleet.ActivityTypeAddedSoftware{}.ActivityName(), fmt.Sprintf(`{"software_title": "ruby", "software_package": "ruby.deb", "team_name": "%s", "team_id": %d}`, createTeamResp.Team.Name, createTeamResp.Team.ID), 0)
-
-		// upload again fails
-		s.uploadSoftwareInstaller(payload, http.StatusConflict, "already exists")
-
-		// download the installer
-		r := s.Do("GET", fmt.Sprintf("/api/latest/fleet/software/%d/package?alt=media", titleID), nil, http.StatusOK, "team_id", fmt.Sprintf("%d", *payload.TeamID))
-		checkDownloadResponse(t, r, payload.Filename)
-
-		// create an orbit host, assign to team and request to download the installer
-		host := createOrbitEnrolledHost(t, "windows", "orbit-host-team", s.ds)
-		require.NoError(t, s.ds.AddHostsToTeam(context.Background(), &createTeamResp.Team.ID, []uint{host.ID}))
-		r = s.Do("POST", "/api/fleet/orbit/software_install/package?alt=media", orbitDownloadSoftwareInstallerRequest{
-			InstallerID:  installerID,
-			OrbitNodeKey: *host.OrbitNodeKey,
-		}, http.StatusOK)
-		checkDownloadResponse(t, r, payload.Filename)
-
-		// delete the installer
-		s.Do("DELETE", fmt.Sprintf("/api/latest/fleet/software/%d/package", titleID), nil, http.StatusNoContent, "team_id", fmt.Sprintf("%d", *payload.TeamID))
-
-		// check activity
-		s.lastActivityOfTypeMatches(fleet.ActivityTypeDeletedSoftware{}.ActivityName(), fmt.Sprintf(`{"software_title": "ruby", "software_package": "ruby.deb", "team_name": "%s", "team_id": %d}`, createTeamResp.Team.Name, createTeamResp.Team.ID), 0)
-	})
-}
-
-func (s *integrationMDMTestSuite) TestSoftwareInstallerNewInstallRequestPlatformValidation() {
-	t := s.T()
-
-	hostsByPlatform := map[string]*fleet.Host{
-		"linux": nil, "darwin": nil, "windows": nil,
-	}
-
-	tm, err := s.ds.NewTeam(context.Background(), &fleet.Team{
-		Name:        t.Name(),
-		Description: "desc",
-	})
-	require.NoError(t, err)
-
-	for platform := range hostsByPlatform {
-		h, err := s.ds.NewHost(context.Background(), &fleet.Host{
-			DetailUpdatedAt: time.Now(),
-			LabelUpdatedAt:  time.Now(),
-			PolicyUpdatedAt: time.Now(),
-			SeenTime:        time.Now().Add(-1 * time.Minute),
-			OsqueryHostID:   ptr.String(t.Name() + uuid.New().String()),
-			NodeKey:         ptr.String(t.Name() + uuid.New().String()),
-			Hostname:        fmt.Sprintf("%sfoo.local", t.Name()),
-			Platform:        platform,
-		})
-		require.NoError(t, err)
-		setOrbitEnrollment(t, h, s.ds)
-
-		err = s.ds.AddHostsToTeam(context.Background(), &tm.ID, []uint{h.ID})
-		require.NoError(t, err)
-
-		hostsByPlatform[platform] = h
-	}
-
-	softwareTitles := map[string]uint{
-		"deb": 0, "msi": 0, "exe": 0, "pkg": 0,
-	}
-
-	for kind := range softwareTitles {
-		// TODO(roberto): we need real binaries for exe, msi and pkg to
-		// perform the API calls.
-		mysql.ExecAdhocSQL(t, s.ds, func(q sqlx.ExtContext) error {
-			ctx := context.Background()
-			installScript := fmt.Sprintf(`echo '%s'`, kind)
-			res, err := q.ExecContext(ctx, `INSERT INTO script_contents (md5_checksum, contents) VALUES (UNHEX(md5(?)), ?)`, installScript, installScript)
-			if err != nil {
-				return err
-			}
-			scriptContentID, _ := res.LastInsertId()
-
-			res, err = q.ExecContext(ctx, `INSERT INTO software_titles (name, source) VALUES ('foo', ?)`, kind)
-			if err != nil {
-				return err
-			}
-			titleID, _ := res.LastInsertId()
-			softwareTitles[kind] = uint(titleID)
-
-			_, err = q.ExecContext(ctx, `
-			INSERT INTO software_installers
-				(title_id, filename, version, install_script_content_id, storage_id, team_id, global_or_team_id, pre_install_query)
-			VALUES
-				(?, ?, ?, ?, unhex(?), ?, ?, ?)`,
-				titleID, fmt.Sprintf("installer.%s", kind), "v1.0.0", scriptContentID, hex.EncodeToString([]byte("test")), tm.ID, tm.ID, "foo")
-			return err
-		})
-	}
-
-	testCases := []struct {
-		platform            string
-		supportedInstallers []string
-	}{
-		{"windows", []string{"exe", "msi"}},
-		{"darwin", []string{"pkg"}},
-		{"linux", []string{"deb"}},
-	}
-
-	for _, tc := range testCases {
-		for platform, host := range hostsByPlatform {
-			for _, kind := range tc.supportedInstallers {
-				wantStatus := http.StatusAccepted
-				if tc.platform != platform {
-					wantStatus = http.StatusBadRequest
-				}
-
-				var resp installSoftwareResponse
-				s.DoJSON("POST", fmt.Sprintf("/api/latest/fleet/hosts/%d/software/install/%d", host.ID, softwareTitles[kind]), nil, wantStatus, &resp)
-			}
-		}
-	}
-}
-
-func (s *integrationMDMTestSuite) TestSoftwareInstallerHostRequests() {
-	t := s.T()
-
-	var createTeamResp teamResponse
-	s.DoJSON("POST", "/api/latest/fleet/teams", &fleet.Team{
-		Name: t.Name(),
-	}, http.StatusOK, &createTeamResp)
-	require.NotZero(t, createTeamResp.Team.ID)
-	teamID := &createTeamResp.Team.ID
-
-	var resp installSoftwareResponse
-	// non-existent host
-	s.DoJSON("POST", "/api/latest/fleet/hosts/1/software/install/1", nil, http.StatusNotFound, &resp)
-
-	// create a host that doesn't have fleetd installed
-	h, err := s.ds.NewHost(context.Background(), &fleet.Host{
-		DetailUpdatedAt: time.Now(),
-		LabelUpdatedAt:  time.Now(),
-		PolicyUpdatedAt: time.Now(),
-		SeenTime:        time.Now().Add(-1 * time.Minute),
-		OsqueryHostID:   ptr.String(t.Name() + uuid.New().String()),
-		NodeKey:         ptr.String(t.Name() + uuid.New().String()),
-		Hostname:        fmt.Sprintf("%sfoo.local", t.Name()),
-		Platform:        "linux",
-	})
-	require.NoError(t, err)
-	err = s.ds.AddHostsToTeam(context.Background(), teamID, []uint{h.ID})
-	require.NoError(t, err)
-
-	// request fails
-	resp = installSoftwareResponse{}
-	s.DoJSON("POST", fmt.Sprintf("/api/latest/fleet/hosts/%d/software/install/1", h.ID), nil, http.StatusUnprocessableEntity, &resp)
-
-	// host installs fleetd
-	setOrbitEnrollment(t, h, s.ds)
-	// TODO(roberto) setOrbitEnrollment is a helper function that silently
-	// sets the team_id to NULL. We need to refactor it to accept a
-	// parameter with an optional team value.
-	err = s.ds.AddHostsToTeam(context.Background(), teamID, []uint{h.ID})
-	require.NoError(t, err)
-
-	// request fails because of non-existent title
-	resp = installSoftwareResponse{}
-	s.DoJSON("POST", fmt.Sprintf("/api/latest/fleet/hosts/%d/software/install/1", h.ID), nil, http.StatusBadRequest, &resp)
-
-	payload := &fleet.UploadSoftwareInstallerPayload{
-		InstallScript:     "another install script",
-		PreInstallQuery:   "another pre install query",
-		PostInstallScript: "another post install script",
-		Filename:          "ruby.deb",
-		Title:             "ruby",
-		TeamID:            teamID,
-	}
-	s.uploadSoftwareInstaller(payload, http.StatusOK, "")
-
-	// install script request succeeds
-	titleID := getSoftwareTitleID(t, s.ds, payload.Title, "deb_packages")
-	resp = installSoftwareResponse{}
-	s.DoJSON("POST", fmt.Sprintf("/api/latest/fleet/hosts/%d/software/install/%d", h.ID, titleID), nil, http.StatusAccepted, &resp)
-
-	// Get the results, should be pending
-	getHostSoftwareResp := getHostSoftwareResponse{}
-	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/hosts/%d/software", h.ID), nil, http.StatusOK, &getHostSoftwareResp)
-	require.Len(t, getHostSoftwareResp.Software, 1)
-	require.NotNil(t, getHostSoftwareResp.Software[0].LastInstall)
-	installUUID := getHostSoftwareResp.Software[0].LastInstall.InstallUUID
-
-	gsirr := getSoftwareInstallResultsResponse{}
-	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/software/install/results/%s", installUUID), nil, http.StatusOK, &gsirr)
-	require.NoError(t, gsirr.Err)
-	require.NotNil(t, gsirr.Results)
-	results := gsirr.Results
-	require.Equal(t, installUUID, results.InstallUUID)
-	require.Equal(t, fleet.SoftwareInstallerPending, results.Status)
-
-	// status is reflected in software title response
-	titleResp := getSoftwareTitleResponse{}
-	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/software/titles/%d", titleID), nil, http.StatusOK, &titleResp, "team_id", strconv.Itoa(int(*teamID)))
-	// TODO: confirm expected behavior of the title response host counts (unspecified)
-	require.Zero(t, titleResp.SoftwareTitle.HostsCount)
-	require.Nil(t, titleResp.SoftwareTitle.CountsUpdatedAt)
-	require.NotNil(t, titleResp.SoftwareTitle.SoftwarePackage)
-	require.Equal(t, "ruby.deb", titleResp.SoftwareTitle.SoftwarePackage.Name)
-	require.NotNil(t, titleResp.SoftwareTitle.SoftwarePackage.Status)
-	require.Equal(t, fleet.SoftwareInstallerStatusSummary{
-		Installed: 0,
-		Pending:   1,
-		Failed:    0,
-	}, *titleResp.SoftwareTitle.SoftwarePackage.Status)
-
-	// status is reflected in list hosts responses and counts when filtering by software title and status
-	var listResp listHostsResponse
-	s.DoJSON("GET", "/api/latest/fleet/hosts", nil, http.StatusOK, &listResp, "software_status", "pending", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Len(t, listResp.Hosts, 1)
-	require.Equal(t, h.ID, listResp.Hosts[0].ID)
-
-	var countResp countHostsResponse
-	s.DoJSON("GET", "/api/latest/fleet/hosts/count", nil, http.StatusOK, &countResp, "software_status", "pending", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Equal(t, 1, countResp.Count)
-
-	listResp = listHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts", nil, http.StatusOK, &listResp, "software_status", "failed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Len(t, listResp.Hosts, 0)
-
-	countResp = countHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts/count", nil, http.StatusOK, &countResp, "software_status", "failed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Equal(t, 0, countResp.Count)
-
-	listResp = listHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts", nil, http.StatusOK, &listResp, "software_status", "installed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Len(t, listResp.Hosts, 0)
-
-	countResp = countHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts/count", nil, http.StatusOK, &countResp, "software_status", "installed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Equal(t, 0, countResp.Count)
-
-	var labelResp createLabelResponse
-	s.DoJSON("POST", "/api/latest/fleet/labels", &createLabelRequest{fleet.LabelPayload{
-		Name:  "test",
-		Hosts: []string{h.Hostname},
-	}}, http.StatusOK, &labelResp)
-	require.NotZero(t, labelResp.Label.ID)
-
-	listResp = listHostsResponse{}
-	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/labels/%d/hosts", labelResp.Label.ID), nil, http.StatusOK, &listResp)
-	require.Len(t, listResp.Hosts, 1)
-	require.Equal(t, h.ID, listResp.Hosts[0].ID)
-
-	countResp = countHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts/count", nil, http.StatusOK, &countResp, "software_status", "pending", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)), "label_id", strconv.Itoa(int(labelResp.Label.ID)))
-	require.Equal(t, 1, countResp.Count)
-
-	listResp = listHostsResponse{}
-	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/labels/%d/hosts", labelResp.Label.ID), nil, http.StatusOK, &listResp, "software_status", "pending", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Len(t, listResp.Hosts, 1)
-	require.Equal(t, h.ID, listResp.Hosts[0].ID)
-
-	countResp = countHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts/count", nil, http.StatusOK, &countResp, "software_status", "pending", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)), "label_id", strconv.Itoa(int(labelResp.Label.ID)))
-	require.Equal(t, 1, countResp.Count)
-
-	listResp = listHostsResponse{}
-	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/labels/%d/hosts", labelResp.Label.ID), nil, http.StatusOK, &listResp, "software_status", "installed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Len(t, listResp.Hosts, 0)
-
-	countResp = countHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts/count", nil, http.StatusOK, &countResp, "software_status", "installed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)), "label_id", strconv.Itoa(int(labelResp.Label.ID)))
-	require.Equal(t, 0, countResp.Count)
-
-	listResp = listHostsResponse{}
-	s.DoJSON("GET", fmt.Sprintf("/api/latest/fleet/labels/%d/hosts", labelResp.Label.ID), nil, http.StatusOK, &listResp, "software_status", "failed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)))
-	require.Len(t, listResp.Hosts, 0)
-
-	countResp = countHostsResponse{}
-	s.DoJSON("GET", "/api/latest/fleet/hosts/count", nil, http.StatusOK, &countResp, "software_status", "failed", "team_id", strconv.Itoa(int(*teamID)), "software_title_id", strconv.Itoa(int(titleID)), "label_id", strconv.Itoa(int(labelResp.Label.ID)))
-	require.Equal(t, 0, countResp.Count)
-
-	// filter validations
-	r := s.Do("GET", "/api/latest/fleet/hosts", nil, http.StatusBadRequest, "software_status", "uninstalled")
-	require.Contains(t, extractServerErrorText(r.Body), "Invalid software_status")
-	r = s.Do("GET", "/api/latest/fleet/hosts", nil, http.StatusBadRequest, "software_status", "installed")
-	require.Contains(t, extractServerErrorText(r.Body), "Missing software_title_id")
-	r = s.Do("GET", "/api/latest/fleet/hosts", nil, http.StatusBadRequest, "software_status", "installed", "software_title_id", "1")
-	require.Contains(t, extractServerErrorText(r.Body), "Missing team_id")
-	r = s.Do("GET", "/api/latest/fleet/hosts", nil, http.StatusBadRequest, "software_status", "installed", "team_id", "1")
-	require.Contains(t, extractServerErrorText(r.Body), "Missing software_title_id")
-	r = s.Do("GET", "/api/latest/fleet/hosts", nil, http.StatusBadRequest, "software_status", "installed", "team_id", "1", "software_title_id", "1", "software_version_id", "1")
-	require.Contains(t, extractServerErrorText(r.Body), "Invalid parameters. The combination of software_version_id and software_title_id is not allowed.")
-	r = s.Do("GET", "/api/latest/fleet/hosts", nil, http.StatusBadRequest, "software_status", "installed", "team_id", "1", "software_title_id", "1", "software_id", "1")
-	require.Contains(t, extractServerErrorText(r.Body), "Invalid parameters. The combination of software_id and software_title_id is not allowed.")
-
-	// TODO(roberto): once we have endpoints to retrieve installers,
-	// request them using the orbit node key
-
-	// TODO(sarah): test other statuses once we have endpoints to set results via orbit
-}
-
-func (s *integrationMDMTestSuite) TestHostSoftwareInstallResult() {
-	ctx := context.Background()
-	t := s.T()
-
-	host := createOrbitEnrolledHost(t, "linux", "", s.ds)
-
-	// create a software installer and some host install requests
-	payload := &fleet.UploadSoftwareInstallerPayload{
-		InstallScript:     "install script",
-		PreInstallQuery:   "pre install query",
-		PostInstallScript: "post install script",
-		Filename:          "ruby.deb",
-		Title:             "ruby",
-	}
-	s.uploadSoftwareInstaller(payload, http.StatusOK, "")
-	titleID := getSoftwareTitleID(t, s.ds, payload.Title, "deb_packages")
-
-	latestInstallUUID := func() string {
-		var id string
-		mysql.ExecAdhocSQL(t, s.ds, func(q sqlx.ExtContext) error {
-			return sqlx.GetContext(ctx, q, &id, `SELECT execution_id FROM host_software_installs ORDER BY id DESC LIMIT 1`)
-		})
-		return id
-	}
-
-	// create some install requests for the host
-	installUUIDs := make([]string, 3)
-	for i := 0; i < len(installUUIDs); i++ {
-		resp := installSoftwareResponse{}
-		s.DoJSON("POST", fmt.Sprintf("/api/v1/fleet/hosts/%d/software/install/%d", host.ID, titleID), nil, http.StatusAccepted, &resp)
-		installUUIDs[i] = latestInstallUUID()
-	}
-
-	type result struct {
-		HostID      uint
-		InstallUUID string
-		Status      fleet.SoftwareInstallerStatus
-	}
-	checkResults := func(want result) {
-		var resp getSoftwareInstallResultsResponse
-		s.DoJSON("GET", "/api/v1/fleet/software/install/results/"+want.InstallUUID, nil, http.StatusOK, &resp)
-
-		assert.Equal(t, want.HostID, resp.Results.HostID)
-		assert.Equal(t, want.InstallUUID, resp.Results.InstallUUID)
-		assert.Equal(t, want.Status, resp.Results.Status)
-	}
-
-	s.Do("POST", "/api/fleet/orbit/software_install/result",
-		json.RawMessage(fmt.Sprintf(`{
-			"orbit_node_key": %q,
-			"install_uuid": %q,
-			"pre_install_condition_output": "1",
-			"install_script_exit_code": 1,
-			"install_script_output": "failed"
-		}`, *host.OrbitNodeKey, installUUIDs[0])),
-		http.StatusNoContent)
-	checkResults(result{
-		HostID:      host.ID,
-		InstallUUID: installUUIDs[0],
-		Status:      fleet.SoftwareInstallerFailed,
-	})
-	wantAct := fleet.ActivityTypeInstalledSoftware{
-		HostID:          host.ID,
-		HostDisplayName: host.DisplayName(),
-		SoftwareTitle:   payload.Title,
-		InstallUUID:     installUUIDs[0],
-		Status:          string(fleet.SoftwareInstallerFailed),
-	}
-	s.lastActivityMatches(wantAct.ActivityName(), string(jsonMustMarshal(t, wantAct)), 0)
-
-	s.Do("POST", "/api/fleet/orbit/software_install/result",
-		json.RawMessage(fmt.Sprintf(`{
-			"orbit_node_key": %q,
-			"install_uuid": %q,
-			"pre_install_condition_output": ""
-		}`, *host.OrbitNodeKey, installUUIDs[1])),
-		http.StatusNoContent)
-	checkResults(result{
-		HostID:      host.ID,
-		InstallUUID: installUUIDs[1],
-		Status:      fleet.SoftwareInstallerFailed,
-	})
-	wantAct.InstallUUID = installUUIDs[1]
-	s.lastActivityOfTypeMatches(wantAct.ActivityName(), string(jsonMustMarshal(t, wantAct)), 0)
-
-	s.Do("POST", "/api/fleet/orbit/software_install/result",
-		json.RawMessage(fmt.Sprintf(`{
-			"orbit_node_key": %q,
-			"install_uuid": %q,
-			"pre_install_condition_output": "1",
-			"install_script_exit_code": 0,
-			"install_script_output": "success",
-			"post_install_script_exit_code": 0,
-			"post_install_script_output": "ok"
-		}`, *host.OrbitNodeKey, installUUIDs[2])),
-		http.StatusNoContent)
-	checkResults(result{
-		HostID:      host.ID,
-		InstallUUID: installUUIDs[2],
-		Status:      fleet.SoftwareInstallerInstalled,
-	})
-	wantAct.InstallUUID = installUUIDs[2]
-	wantAct.Status = string(fleet.SoftwareInstallerInstalled)
-	lastActID := s.lastActivityOfTypeMatches(wantAct.ActivityName(), string(jsonMustMarshal(t, wantAct)), 0)
-
-	// non-existing installation uuid
-	s.Do("POST", "/api/fleet/orbit/software_install/result",
-		json.RawMessage(fmt.Sprintf(`{
-			"orbit_node_key": %q,
-			"install_uuid": "uuid-no-such",
-			"pre_install_condition_output": ""
-		}`, *host.OrbitNodeKey)),
-		http.StatusNotFound)
-	// no new activity created
-	s.lastActivityOfTypeMatches(wantAct.ActivityName(), string(jsonMustMarshal(t, wantAct)), lastActID)
-}
-
-func (s *integrationMDMTestSuite) uploadSoftwareInstaller(payload *fleet.UploadSoftwareInstallerPayload, expectedStatus int, expectedError string) {
-	t := s.T()
-	openFile := func(name string) *os.File {
-		f, err := os.Open(filepath.Join("testdata", "software-installers", name))
-		require.NoError(t, err)
-		return f
-	}
-
-	f := openFile(payload.Filename)
-	defer f.Close()
-
-	payload.InstallerFile = f
-
-	var b bytes.Buffer
-	w := multipart.NewWriter(&b)
-
-	// add the software field
-	fw, err := w.CreateFormFile("software", payload.Filename)
-	require.NoError(t, err)
-	n, err := io.Copy(fw, payload.InstallerFile)
-	require.NoError(t, err)
-	require.NotZero(t, n)
-
-	// add the team_id field
-	if payload.TeamID != nil {
-		require.NoError(t, w.WriteField("team_id", fmt.Sprintf("%d", *payload.TeamID)))
-	}
-	// add the remaining fields
-	require.NoError(t, w.WriteField("install_script", payload.InstallScript))
-	require.NoError(t, w.WriteField("pre_install_query", payload.PreInstallQuery))
-	require.NoError(t, w.WriteField("post_install_script", payload.PostInstallScript))
-
-	w.Close()
-
-	headers := map[string]string{
-		"Content-Type":  w.FormDataContentType(),
-		"Accept":        "application/json",
-		"Authorization": fmt.Sprintf("Bearer %s", s.token),
-	}
-
-	r := s.DoRawWithHeaders("POST", "/api/latest/fleet/software/package", b.Bytes(), expectedStatus, headers)
-	if expectedError != "" {
-		errMsg := extractServerErrorText(r.Body)
-		require.Contains(t, errMsg, expectedError)
-	}
-}
-
-func getSoftwareTitleID(t *testing.T, ds *mysql.Datastore, title, source string) uint {
-	var id uint
-	mysql.ExecAdhocSQL(t, ds, func(q sqlx.ExtContext) error {
-		return sqlx.GetContext(context.Background(), q, &id, `SELECT id FROM software_titles WHERE name = ? AND source = ? AND browser = ''`, title, source)
-	})
-	return id
 }
